Major overhaul
* Remove redundant code from PoC NVD to OSV converter
* Move repository inference functionality over to CPE Dictionary parser
* Cover an edge case where bare repository URLs have a trailing slash
