+++ vulnfeeds/cmd/cperepos/main.go
@@ -13,6 +13,9 @@ The flags are:
 	  --cpe_dictionary
 		The path to the uncompressed NVD CPE Dictionary XML file, see https://nvd.nist.gov/products/cpe
 
+	  --debian_metadata_path
+	        The path to a directory containing a local mirror of Debian copyright metadata, see README.md
+
 	  --output_dir
 	        The directory to output cpe_product_to_repo.json and cpe_reference_description_frequency.csv in
 
@@ -25,6 +28,8 @@ The flags are:
 package main
 
 import (
+	"bufio"
+	"bytes"
 	"context"
 	"encoding/csv"
 	"encoding/json"
@@ -34,12 +39,15 @@ import (
 	"io"
 	"io/ioutil"
 	"log"
+	"net/url"
 	"os"
+	"path"
 	"path/filepath"
 	"regexp"
 	"sort"
 	"strconv"
 	"strings"
+	"text/template"
 
 	"cloud.google.com/go/logging"
 	"github.com/google/osv/vulnfeeds/cves"
@@ -68,6 +76,8 @@ type Reference struct {
 type CPE23Item struct {
 	Name string `xml:"name,attr"`
 }
+
+type VendorProductToRepoMap map[string]map[string][]string
 
 const (
 	CPEDictionaryDefault = "cve_jsons/nvdcpematch-1.0.json"
@@ -114,11 +124,12 @@ var (
 		"https://github.com/z00z00z00/Safenet_SAC_CVE-2021-42056",
 	}
 	// Match repos with "CVE", "CVEs" or a pure CVE number in their name, anything from GitHubAssessments
-	InvalidRepoRegex  = `/(?:(?:CVEs?)|(?:CVE-\d{4}-\d{4,})|GitHubAssessments/.*|advisories/GHSA.*)$`
-	CPEDictionaryFile = flag.String("cpe_dictionary", CPEDictionaryDefault, "CPE Dictionary file to parse")
-	OutputDir         = flag.String("output_dir", OutputDirDefault, "Directory to output cpe_product_to_repo.json and cpe_reference_description_frequency.csv in")
-	GCPLoggingProject = flag.String("gcp_logging_project", projectId, "GCP project ID to use for logging, set to an empty string to log locally only")
-	Verbose           = flag.Bool("verbose", false, "Output some telemetry to stdout during execution")
+	InvalidRepoRegex   = `/(?:(?:CVEs?)|(?:CVE-\d{4}-\d{4,})|GitHubAssessments/.*|advisories/GHSA.*)$`
+	CPEDictionaryFile  = flag.String("cpe_dictionary", CPEDictionaryDefault, "CPE Dictionary file to parse")
+	OutputDir          = flag.String("output_dir", OutputDirDefault, "Directory to output cpe_product_to_repo.json and cpe_reference_description_frequency.csv in")
+	GCPLoggingProject  = flag.String("gcp_logging_project", projectId, "GCP project ID to use for logging, set to an empty string to log locally only")
+	DebianMetadataPath = flag.String("debian_metadata_path", "", "Path to Debian copyright metadata")
+	Verbose            = flag.Bool("verbose", false, "Output some telemetry to stdout during execution")
 )
 
 func LoadCPEDictionary(f string) (CPEDict, error) {
@@ -138,7 +149,7 @@ func LoadCPEDictionary(f string) (CPEDict, error) {
 }
 
 // Outputs a JSON file of the product-to-repo map.
-func outputProductToRepoMap(prm map[string][]string, f io.Writer) error {
+func outputProductToRepoMap(prm VendorProductToRepoMap, f io.Writer) error {
 	productsWithoutRepos := 0
 	for p := range prm {
 		if len(prm[p]) == 0 {
@@ -195,11 +206,118 @@ func outputDescriptionFrequency(df map[string]int, f io.Writer) error {
 	return nil
 }
 
+// Checks if a URL relates to the FSF.
+func IsGNUURL(url string) bool {
+	re := regexp.MustCompile(`^https?://.*\.(?:non)?gnu\.org/`)
+
+	return re.MatchString(url)
+}
+
+// Tries to translate Savannah URLs to their corresponding Git repository URL.
+func MaybeTranslateSavannahURL(u string) (string, bool) {
+	type GNUPlaceholder struct {
+		GNUOrNonGNU string
+		Path        string
+	}
+	var tpl bytes.Buffer
+
+	supportedHostnames := []string{
+		"download.savannah.gnu.org",
+		"savannah.gnu.org",
+		"download.savannah.gnu.org",
+		"download-mirror.savannah.gnu.org",
+		"download.savannah.nongnu.org",
+		"savannah.nongnu.org",
+		"download.savannah.nongnu.org",
+		"download-mirror.savannah.nongnu.org",
+	}
+
+	// Get hostname out of URL
+	parsedURL, err := url.Parse(u)
+	if err != nil {
+		panic(err)
+	}
+
+	if slices.Contains(supportedHostnames, parsedURL.Hostname()) {
+		hostnameParts := strings.Split(parsedURL.Hostname(), ".")
+		// Pull out the "nongnu" or "gnu" part of the hostname
+		domain := GNUPlaceholder{hostnameParts[len(hostnameParts)-2], path.Base(parsedURL.Path)}
+		savannahGitRepoTemplate, err := template.New("SavannahGitRepoURL").Parse("https://git.savannah.{{ .GNUOrNonGNU }}.org/git/{{ .Path }}.git")
+		if err != nil {
+			panic(err)
+		}
+		err = savannahGitRepoTemplate.Execute(&tpl, domain)
+		return tpl.String(), true
+	}
+
+	// Return the original URL unmodified
+	return u, false
+}
+
+// Returns the data associated with the ^Source: line of a machine-readable Debian copyright file
+// See https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/
+func MaybeGetSourceFromDebianCopyright(copyrightFile string) (string, bool) {
+	re := regexp.MustCompile(`^Source: (.*)$`)
+
+	file, err := os.Open(copyrightFile)
+	if err != nil {
+		Logger.Fatalf("%v", err)
+	}
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+
+	// Check the first line to see if we have a machine-readable copyright file
+	scanner.Scan()
+	if scanner.Text() != "Format: https://www.debian.org/doc/packaging-manuals/copyright-format/1.0/" {
+		return "", false
+	}
+
+	for scanner.Scan() {
+		line := scanner.Text()
+		match := re.FindStringSubmatch(line)
+		if match != nil {
+			return match[1], true
+		}
+	}
+
+	return "", false
+}
+
+// Tries to find a Debian copyright file for the package and returns the source URL if IsRepoURL() agrees.
+func MaybeGetSourceRepoFromDebian(mdir string, pkg string) string {
+	var metadata string
+	if strings.HasPrefix(pkg, "lib") {
+		metadata = path.Join(mdir, "changelogs/main", string(pkg[0:4]), pkg, "unstable_copyright")
+	} else {
+		metadata = path.Join(mdir, "changelogs/main", string(pkg[0]), pkg, "unstable_copyright")
+	}
+	if _, err := os.Stat(metadata); err == nil {
+		// parse the copyright file and go from here
+		Logger.Infof("FYI: Will look at %s", metadata)
+		possibleRepo, ok := MaybeGetSourceFromDebianCopyright(metadata)
+		if !ok {
+			return ""
+		}
+		if utility.IsRepoURL(possibleRepo) {
+			return possibleRepo
+		}
+		// Incorporate Savannah URL to Git translation here
+		if IsGNUURL(possibleRepo) {
+			repo, translated := MaybeTranslateSavannahURL(possibleRepo)
+			if translated {
+				return repo
+			}
+		}
+		Logger.Infof("FYI: Disregarding %s", possibleRepo)
+	}
+	return ""
+}
+
 // Analyze CPE Dictionary
-func analyzeCPEDictionary(d CPEDict) (ProductToRepo map[string][]string, DescriptionFrequency map[string]int, ProductToVendor map[string][]string) {
-	ProductToRepo = make(map[string][]string)
+func analyzeCPEDictionary(d CPEDict) (ProductToRepo VendorProductToRepoMap, DescriptionFrequency map[string]int) {
+	ProductToRepo = make(VendorProductToRepoMap)
 	DescriptionFrequency = make(map[string]int)
-	ProductToVendor = make(map[string][]string)
 	for _, c := range d.CPEItems {
 		CPE, err := cves.ParseCPE(c.CPE23.Name)
 		if err != nil {
@@ -210,41 +328,40 @@ func analyzeCPEDictionary(d CPEDict) (ProductToRepo map[string][]string, Descrip
 			// Not interested in hardware or operating systems.
 			continue
 		}
-		// Gather the data to answer the question: "are there product name collisions?"
-		if *Verbose {
-			if _, exists := ProductToVendor[CPE.Product]; exists {
-				if !slices.Contains(ProductToVendor[CPE.Product], CPE.Vendor) {
-					ProductToVendor[CPE.Product] = append(ProductToVendor[CPE.Product], CPE.Vendor)
-				}
-			} else {
-				ProductToVendor[CPE.Product] = []string{CPE.Vendor}
-			}
-		}
 		for _, r := range c.References {
 			DescriptionFrequency[r.Description] += 1
 			repo, err := cves.Repo(r.URL)
 			if err != nil {
-				Logger.Infof("Disregarding %q for %q (%s) because %v", r.URL, CPE.Product, r.Description, err)
+				Logger.Infof("Disregarding %q for %q/%q (%s) because %v", r.URL, CPE.Vendor, CPE.Product, r.Description, err)
 				continue
 			}
 			// Disregard the repos we know we don't like.
 			matched, _ := regexp.MatchString(InvalidRepoRegex, repo)
 			if matched {
-				Logger.Infof("Disliking %q for %q (%s) (matched regexp)", repo, CPE.Product, r.Description)
+				Logger.Infof("Disliking %q for %q/%q (%s) (matched regexp)", repo, CPE.Vendor, CPE.Product, r.Description)
 				continue
 			}
 			if slices.Contains(InvalidRepos, repo) {
-				Logger.Infof("Disliking %q for %q (%s)", repo, CPE.Product, r.Description)
+				Logger.Infof("Disliking %q for %q/%q (%s)", repo, CPE.Vendor, CPE.Product, r.Description)
 				continue
 			}
-			if slices.Contains(ProductToRepo[CPE.Product], repo) {
+			if slices.Contains(ProductToRepo[CPE.Vendor][CPE.Product], repo) {
 				continue
 			}
-			Logger.Infof("Liking %q for %q (%s)", repo, CPE.Product, r.Description)
-			ProductToRepo[CPE.Product] = append(ProductToRepo[CPE.Product], repo)
+			Logger.Infof("Liking %q for %q/%q (%s)", repo, CPE.Vendor, CPE.Product, r.Description)
+			ProductToRepo[CPE.Vendor] = map[string][]string{CPE.Product: append(ProductToRepo[CPE.Vendor][CPE.Product], repo)}
+		}
+		// If we've arrived to this point and not calculated any repos for the product, try Debian
+		if len(ProductToRepo[CPE.Vendor][CPE.Product]) == 0 && *DebianMetadataPath != "" {
+			Logger.Infof("Trying to derive a repo from Debian for %q/%q", CPE.Vendor, CPE.Product)
+			repo := MaybeGetSourceRepoFromDebian(*DebianMetadataPath, CPE.Product)
+			if repo != "" {
+				Logger.Infof("Derived repo: %s", repo)
+				ProductToRepo[CPE.Vendor] = map[string][]string{CPE.Product: []string{repo}}
+			}
 		}
 	}
-	return ProductToRepo, DescriptionFrequency, ProductToVendor
+	return ProductToRepo, DescriptionFrequency
 }
 
 func main() {
@@ -269,7 +386,7 @@ func main() {
 		Logger.Fatalf("Failed to load %s: %v", *CPEDictionaryFile, err)
 	}
 
-	productToRepo, descriptionFrequency, productToVendor := analyzeCPEDictionary(CPEDictionary)
+	productToRepo, descriptionFrequency := analyzeCPEDictionary(CPEDictionary)
 
 	mappingFile, err := os.Create(filepath.Join(*OutputDir, "cpe_product_to_repo.json"))
 	if err != nil {
@@ -279,16 +396,6 @@ func main() {
 	err = outputProductToRepoMap(productToRepo, mappingFile)
 	if err != nil {
 		Logger.Fatalf("%v", err)
-	}
-	// Answer the question: "are there product name collisions?"
-	if *Verbose {
-		for product, vendors := range productToVendor {
-			if len(vendors) == 1 {
-				continue
-			}
-			Logger.Infof("Product %s has >1 vendors: [%s]", product, strings.Join(vendors, ", "))
-			fmt.Printf("Product %s has >1 vendors: [%s]\n", product, strings.Join(vendors, ", "))
-		}
 	}
 	frequencyFile, err := os.Create(filepath.Join(*OutputDir, "cpe_reference_description_frequency.csv"))
 	if err != nil {
