+++ vulnfeeds/utility/utility.go
@@ -1,5 +1,7 @@
 package utility
 
+import "regexp"
+
 // SliceEqual returns true if two slices have identical items in the same order
 func SliceEqual[K comparable](a []K, b []K) bool {
 	if len(a) != len(b) {
@@ -31,4 +33,11 @@ func SliceEqualUnordered[K comparable](a []K, b []K) bool {
 		}
 	}
 	return true
+}
+
+// Checks if a URL is to a supported repo.
+func IsRepoURL(url string) bool {
+	re := regexp.MustCompile(`http[s]?:\/\/(?:c?git(?:hub|lab)?)\.|\.git$`)
+
+	return re.MatchString(url)
 }
